+++ Credixware/Internal/Hacks/Skinchanger.h
@@ -8,7 +8,36 @@
 bool bKnifeEquipped = false;
 int iLocalID = 0;
 
-namespace Skinchanger {
+class Weapon {
+public:
+	void SetItemDefinition(int iValue) {
+		*(int*)(this + Offsets::m_iItemDefinitionIndex) = iValue;
+	}
+	void SetItemIDHigh(int iValue) {
+		*(int*)(this + Offsets::m_iItemIDHigh) = iValue;
+	}
+	void SetPaintKit(int iValue) {
+		*(int*)(this + Offsets::m_nFallbackPaintKit) = iValue;
+	}
+	void SetSeed(int iValue) {
+		*(int*)(this + Offsets::m_nFallbackSeed) = iValue;
+	}
+
+	int GetItemDefinition() {
+		return *(int*)(this + Offsets::m_iItemDefinitionIndex);
+	}
+	int GetItemIDHigh() {
+		return *(int*)(this + Offsets::m_iItemIDHigh);
+	}
+	int GetPaintKit() {
+		return *(int*)(this + Offsets::m_nFallbackPaintKit);
+	}
+	int GetSeed() {
+		return *(int*)(this + Offsets::m_nFallbackSeed);
+	}
+};
+
+namespace SkinChanger {
 	void Init() {
 		player_info_t localPlayer;
 
@@ -16,20 +45,13 @@ namespace Skinchanger {
 			iLocalID = localPlayer.user_id;
 		}
 	}
-	void Update() {
-		/*IClientEntity* LocalPlayer = g_pClientEntityList->GetClientEntity(g_pEngineClient->GetLocalPlayer());
-		UINT* hWeapons = (UINT*)((DWORD)LocalPlayer->GetBaseEntity() + Offsets::m_hMyWeapons);
-		for (int i = 0; hWeapons[i]; i++) {
-			if (hWeapons[i] > 0) {
-				printf("%i: %i\n", hWeapons[i], hWeapons[i]+Offsets::m_iItemDefinitionIndex);
-			}
-		}*/
+	void Update() {	
 		if (Settings::SkinChanger::bKnifeChangerEnabled && Settings::SkinChanger::Knife != 0) {
 			for (int i = 0; i < g_pClientEntityList->GetHighestEntityIndex(); i++) {
 				IClientEntity* Entity = g_pClientEntityList->GetClientEntity(i);
 				if (Entity) {
 					ClientClass* EntityClass = Entity->GetClientClass();
-					if (strstr(EntityClass->GetName(), "CPredictedViewModel")) {
+					if (EntityClass->m_ClassID == CPREDICTEDVIEWMODEL) {
 						C_BaseEntity* BaseEntity = Entity->GetBaseEntity();
 						if (BaseEntity && bKnifeEquipped) {
 							int karam = g_pModelInfo->GetModelIndex(KnifeModels[Settings::SkinChanger::Knife].c_str());